+++ DemoApp/Src/TerrainVoxel.h
@@ -1,5 +1,5 @@
 /**
- * @version     2.0.0 23-Dec-14
+ * @version     2.1.0 01-Jan-15
  * @copyright   Copyright (c) 2014-2015 by Andy Liebke. All rights reserved.
  */
 #ifndef __TERRAIN_VOXEL_H__
@@ -17,7 +17,7 @@ class TerrainVoxel : public TerrainAbstract
         TerrainVoxel& operator = (const TerrainVoxel& src);
     
         void setVoxelState(const char state, const unsigned short x, const unsigned short y, const unsigned short z);
-        void setGridNode(const unsigned short x, const unsigned short y, const unsigned short z);
+        void setGridNode(const unsigned short x, const float y, const unsigned short z);
         void render(void);
         void release(void);
